Benchmark using FST serialization is about 3x faster than PB on reads

+++ es74x/src/main/scala/com/klibisz/elastiknn/query/ExactQuery.scala
@@ -26,7 +26,7 @@ object ExactQuery {
             val binaryValue = vecDocVals.binaryValue()
             codec.decode(binaryValue.bytes.take(binaryValue.length))
           } else throw new RuntimeException(s"Couldn't advance to doc with id [$docId]")
-          simFunc(queryVec, storedVec).toFloat
+          simFunc(queryVec, storedVec)
         }
 
         override def explainScore(docId: Int, subQueryScore: Explanation): Explanation =
